requirements.txt
streamlit
langchain
langchain_openai

python-dotenv
